added rounting logics and refactored the code
