+++ src/components/CheckListPage.jsx
@@ -1,0 +1,12 @@
+/* eslint-disable react/jsx-props-no-spreading */
+import React from 'react';
+
+function CheckList() {
+  return (
+    <div className="checklist">
+      <h2>This is the list of all items on that specific checklist</h2>
+    </div>
+  );
+}
+
+export default CheckList;
