+++ src/components/EditableControls.jsx
@@ -1,0 +1,28 @@
+/* eslint-disable react/jsx-props-no-spreading */
+import React from 'react';
+import {
+  Flex, ButtonGroup, IconButton, useEditableControls,
+} from '@chakra-ui/react';
+import { CheckIcon, CloseIcon, EditIcon } from '@chakra-ui/icons';
+
+function EditableControls() {
+  const {
+    isEditing,
+    getSubmitButtonProps,
+    getCancelButtonProps,
+    getEditButtonProps,
+  } = useEditableControls();
+
+  return isEditing ? (
+    <ButtonGroup justifyContent="center" size="sm">
+      <IconButton icon={<CheckIcon />} {...getSubmitButtonProps()} />
+      <IconButton icon={<CloseIcon />} {...getCancelButtonProps()} />
+    </ButtonGroup>
+  ) : (
+    <Flex justifyContent="center">
+      <IconButton size="sm" icon={<EditIcon />} {...getEditButtonProps()} />
+    </Flex>
+  );
+}
+
+export default EditableControls;
