+++ src/components/Test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function TestFunc() {
+  return (
+    <h2>Test Page</h2>
+  );
+}
+
+export default TestFunc;
