+++ src/components/VolumeDemo.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import {
+  Box,
+  Slider,
+  SliderTrack,
+  SliderFilledTrack,
+  SliderThumb,
+} from '@chakra-ui/react';
+
+function VolumeDemo() {
+  return (
+    <Slider defaultValue={60} min={0} max={300} step={30}>
+      <SliderTrack bg="red.100">
+        <Box position="relative" right={10} />
+        <SliderFilledTrack bg="tomato" />
+      </SliderTrack>
+      <SliderThumb boxSize={6} />
+    </Slider>
+  );
+}
+
+export default VolumeDemo;
